Tue, Feb 7, 2023, 12:05 AM +05:30

+++ cs50p/bitcoin/bitcoin.py
@@ -4,7 +4,7 @@ if len(sys.argv) != 2:
     sys.exit("Missing command-line argument ")
 
 try:
-    coin = int(sys.argv[1])
+    coin = float(sys.argv[1])
 except ValueError:
     sys.exit("Command-line argument is not a number")
